Updated Code
Added error messages if the number of inflows do not match.

The model is mass conservative if inflow hydrographs are entered away from the borders.

Instability issues arise when the water reach the borders of the domain, suggesting that the model is calculating some water that is leaving the system.

This is something that @luiscastillo1993  and @mateohs23  could check.

Put an inflow hydrograph at the middle of the domain and check the mass balance. Then, later, put the same hydrograph in a few cells of the border and check the hydrograph.

I feel we need to implement a proper better boundary condition at the borders of the domain to make sure any water if flowing outside of it.